requirements.txt
numpy==1.24.3
openai==1.68.2
pandas==2.0.3
python_igraph==0.11.8
scikit_learn==1.3.0
scipy==1.15.2
torch==2.2.1
transformers==4.32.1
